requirements.txt
PyYAML>=6.0
Jinja2>=3.1
pytest>=7.0
flask>=3.0
arxiv>=2.0
